Create skeleton app
Initial commit of scaffolding using vue and Node-express.  I elected
Vue for the front end because it's a framework I need to learn for
other projects.  i18n has been added just so I can get some practice
with WCAG requirements.  The server has an api route example, as well
as serves the VUE app out of the public subfolder of the server
folder.  This is where the Vue app builds to in production.
Readme will be expanded upon as we find out what is needed to get
other people up and running with minimal issue.
All current Vue components are examples, and will likely be removed
shortly once actual components start getting built.  Folder
structure is also likely to change as the project gets underway
and we learn more about the framework.