Add an api route with login skeleton
Authorization framework will be using JWT.  This commit creates a
login endpoint that returns a JWT token.  This token can be used
to access a secure api route.  Still need to figure out how to store
the token securely on the client side.  Current standard practice
seems to be splitting the jwt into 2 cookies.  This will be
implemented in a future commit.