Use jwt tokens for authentication
Set up jwt cookies for authentication, and a login api to pass the
token to the client.  HttpOnly is used for the jwt so that the client
can't read it's contents.  It should also use the Secure option, but
I don't have https in set up in my dev environment and that would
break things.  The login component will accept anything at the moment.
Authenticating the user name and password is future work.
Expiration time on authentication is set to 15 minutes for now.  This
should probably be shortened and refreshed on any API interaction in
the future.  The login component is not styled yet.